moved new game store
